requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
astroid==3.2.2
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
cfgv==3.4.0
click==8.1.7
dataclasses-json==0.6.6
deepgram-sdk==3.2.7
dill==0.3.8
distlib==0.3.8
distro==1.9.0
filelock==3.14.0
flake8==7.0.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.36
idna==3.7
iniconfig==2.0.0
isort==5.13.2
marshmallow==3.21.2
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
openai==1.30.1
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
pycodestyle==2.11.1
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pyflakes==3.2.0
pylint==3.2.2
pytest==8.2.1
pytest-asyncio==0.23.7
python-dotenv==1.0.1
PyYAML==6.0.1
setuptools==70.0.0
sniffio==1.3.1
tomlkit==0.12.5
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
verboselogs==1.7
virtualenv==20.26.2
websockets==12.0
yarl==1.9.4
